Make ProductVariant a base variant class; add SKU and price fields in Product
