+++ saleor/product/models/__init__.py
@@ -1,7 +1,7 @@
 from .base import Category, Product
 from .discounts import FixedProductDiscount, get_product_discounts
-from .products import (GenericProduct, GenericVariant, PhysicalProduct)
-from .variants import (StockedProduct, ProductVariant)
+from .products import (GenericProduct, GenericVariant, PhysicalProduct,
+                       ProductVariant, StockedProduct)
 from .images import ProductImage
 
 __all__ = ['Category', 'Product', 'FixedProductDiscount',
