+++ saleor/product/models/products.py
@@ -1,19 +1,16 @@
 from __future__ import unicode_literals
+from decimal import Decimal
 
-from django.conf import settings
+from django.core.validators import MinValueValidator
 from django.db import models
 from django.utils.translation import pgettext_lazy
-from django_prices.models import PriceField
+from satchless.item import StockedItem
 
-from .base import Product
-from .variants import (ProductVariant, StockedProduct)
+from .base import Product, ProductVariant
 
 
 class PhysicalProduct(models.Model):
     weight = models.DecimalField(max_digits=6, decimal_places=2)
-    price = PriceField(
-        pgettext_lazy('Product field', 'price'),
-        currency=settings.DEFAULT_CURRENCY, max_digits=12, decimal_places=4)
 
     class Meta:
         abstract = True
@@ -25,23 +22,30 @@ class PhysicalProduct(models.Model):
         except AttributeError:
             return self.product.weight
 
-    def is_available(self):
-        if self.variants.exists():
-            return any(variant.is_available() for variant in self)
-        else:
-            return False
+
+class StockedProduct(models.Model, StockedItem):
+    stock = models.IntegerField(pgettext_lazy('Product item field', 'stock'),
+                                validators=[MinValueValidator(0)],
+                                default=Decimal(1))
+
+    class Meta:
+        abstract = True
+        app_label = 'product'
+
+    def get_stock(self):
+        return self.stock
+
+    def is_item_available(self):
+        return self.get_stock() > 0
 
 
-class GenericProduct(PhysicalProduct, Product):
+class GenericProduct(StockedProduct, PhysicalProduct, Product):
 
     class Meta:
         app_label = 'product'
 
 
-class GenericVariant(ProductVariant, StockedProduct):
-    product = models.ForeignKey(GenericProduct, related_name='variants')
-    name = models.CharField(pgettext_lazy('Variant field', 'name'),
-                            max_length=100)
+class GenericVariant(StockedProduct, ProductVariant):
 
     class Meta:
         app_label = 'product'
